refactor: remove references of broadcastRawTransaction, use only broadcastTransaction
