+++ src/app/components/fees-row/fees-row.tsx
@@ -63,14 +63,17 @@ export function FeesRow({
   }, [feeHelper, defaultFeeValue, feeTypeHelper]);
 
   useEffect(() => {
+    if (isSponsored) {
+      void feeHelper.setValue(0);
+      return;
+    }
+
     if (!defaultFeeValue && hasFeeEstimates && !feeField.value && !isCustom) {
       void feeHelper.setValue(
         convertAmountToBaseUnit(fees.estimates[FeeTypes.Middle].fee).toString()
       );
       void feeTypeHelper.setValue(FeeTypes[FeeTypes.Middle]);
-    }
-    if (isSponsored) {
-      void feeHelper.setValue(0);
+      return;
     }
   }, [
     defaultFeeValue,
