+++ src/app/pages/transaction-request/transaction-request.tsx
@@ -1,9 +1,8 @@
 import { memo } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
 
-import { Formik } from 'formik';
+import { Formik, FormikHelpers } from 'formik';
 import { Flex } from 'leather-styles/jsx';
-import get from 'lodash.get';
 import * as yup from 'yup';
 
 import { HIGH_FEE_WARNING_LEARN_MORE_URL_STX } from '@shared/constants';
@@ -13,7 +12,6 @@ import { StacksTransactionFormValues } from '@shared/models/form.model';
 import { RouteUrls } from '@shared/route-urls';
 
 import { useAnalytics } from '@app/common/hooks/analytics/use-analytics';
-import { LoadingKeys, useLoading } from '@app/common/hooks/use-loading';
 import { useOnMount } from '@app/common/hooks/use-on-mount';
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
 import { stxFeeValidator } from '@app/common/validation/forms/fee-validators';
@@ -25,6 +23,7 @@ import { HighFeeDrawer } from '@app/features/high-fee-drawer/high-fee-drawer';
 import { ContractCallDetails } from '@app/features/stacks-transaction-request/contract-call-details/contract-call-details';
 import { ContractDeployDetails } from '@app/features/stacks-transaction-request/contract-deploy-details/contract-deploy-details';
 import { FeeForm } from '@app/features/stacks-transaction-request/fee-form';
+import { useStacksBroadcastTransaction } from '@app/features/stacks-transaction-request/hooks/use-stacks-broadcast-transaction';
 import { MinimalErrorMessage } from '@app/features/stacks-transaction-request/minimal-error-message';
 import { PageTop } from '@app/features/stacks-transaction-request/page-top';
 import { PostConditionModeWarning } from '@app/features/stacks-transaction-request/post-condition-mode-warning';
@@ -38,16 +37,12 @@ import { useNextNonce } from '@app/query/stacks/nonce/account-nonces.hooks';
 import { useTransactionRequestState } from '@app/store/transactions/requests.hooks';
 import {
   useGenerateUnsignedStacksTransaction,
-  useSignStacksTransaction,
-  useStacksTransactionBroadcast,
   useUnsignedStacksTransactionBaseState,
 } from '@app/store/transactions/transaction.hooks';
 import { Link } from '@app/ui/components/link/link';
 
 function TransactionRequestBase() {
   const transactionRequest = useTransactionRequestState();
-  const { setIsLoading, setIsIdle } = useLoading(LoadingKeys.SUBMIT_TRANSACTION_REQUEST);
-  // const handleBroadcastTransaction = useSoftwareWalletTransactionRequestBroadcast();
   const unsignedTx = useUnsignedStacksTransactionBaseState();
   const { data: stxFees } = useCalculateStacksTxFees(unsignedTx.transaction);
   const analytics = useAnalytics();
@@ -55,39 +50,31 @@ function TransactionRequestBase() {
   const { data: stacksBalances } = useCurrentStacksAccountBalances();
   const { data: nextNonce } = useNextNonce();
   const navigate = useNavigate();
-  const signStacksTransaction = useSignStacksTransaction();
-  const txBroadcast = useStacksTransactionBroadcast();
+  const { stacksBroadcastTransaction } = useStacksBroadcastTransaction('STX');
 
   useRouteHeader(<PopupHeader />);
 
   useOnMount(() => void analytics.track('view_transaction_signing'));
 
-  const onSubmit = async (values: StacksTransactionFormValues) => {
+  async function onSubmit(
+    values: StacksTransactionFormValues,
+    formikHelpers: FormikHelpers<StacksTransactionFormValues>
+  ) {
+    formikHelpers.setSubmitting(true);
     const unsignedTx = await generateUnsignedTx(values);
 
     if (!unsignedTx)
       return logger.error('Failed to generate unsigned transaction in transaction-request');
 
-    const signedTx = await signStacksTransaction(unsignedTx);
-
-    if (!signedTx) return logger.error('Failed to sign transaction in transaction-request');
-
-    setIsLoading();
-
-    try {
-      await txBroadcast({ signedTx });
-      setIsIdle();
-    } catch (e) {
-      navigate(RouteUrls.TransactionBroadcastError, { state: { message: get(e, 'message') } });
-      return;
-    }
+    await stacksBroadcastTransaction(unsignedTx);
 
     void analytics.track('submit_fee_for_transaction', {
       calculation: stxFees?.calculation,
       fee: values.fee,
       type: values.feeType,
     });
-  };
+    formikHelpers.setSubmitting(false);
+  }
 
   if (!transactionRequest) return null;
