+++ src/app/store/transactions/transaction.ts
@@ -1,20 +1,8 @@
-import { AuthType, ChainID, StacksTransaction, TransactionVersion } from '@stacks/transactions';
+import { ChainID, TransactionVersion } from '@stacks/transactions';
 import { atom } from 'jotai';
 
-import { stacksTransactionToHex } from '@app/common/transactions/stacks/transaction.utils';
 import { whenStacksChainId } from '@app/common/utils';
 import { currentNetworkAtom } from '@app/store/networks/networks';
-
-export function prepareTxDetailsForBroadcast(tx: StacksTransaction) {
-  const serialized = tx.serialize();
-  const txRaw = stacksTransactionToHex(tx);
-
-  return {
-    serialized,
-    isSponsored: tx.auth?.authType === AuthType.Sponsored,
-    txRaw,
-  };
-}
 
 export const transactionNetworkVersionState = atom(get => {
   const currentNetwork = get(currentNetworkAtom);
